bin/requirements.txt
typer
python-dotenv
